backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
fpdf2==2.7.6